- Separated implementations from interfaces, which are now being referenced by higher-layer components instead.
