+++ app/src/main/java/com/pmatuki/wowpapers/remote/api/WallpaperDataSource.kt
@@ -1,0 +1,9 @@
+package com.pmatuki.wowpapers.remote.api
+
+import com.pmatuki.wowpapers.remote.model.WallpaperList
+
+interface WallpaperDataSource {
+
+    suspend fun getWallpapers() : WallpaperList
+
+}
