+++ app/src/main/java/com/pmatuki/wowpapers/view/di/ToothpickScopes.kt
@@ -1,10 +1,12 @@
 package com.pmatuki.wowpapers.view.di
 
-import com.pmatuki.wowpapers.WowpaperApplication
 import com.pmatuki.wowpapers.core.WallpaperService
+import com.pmatuki.wowpapers.core.WallpaperServiceImpl
 import com.pmatuki.wowpapers.di.AppScope
-import com.pmatuki.wowpapers.remote.WallpaperDataSource
+import com.pmatuki.wowpapers.remote.api.WallpaperDataSource
+import com.pmatuki.wowpapers.remote.api.WallpaperDataSourceImpl
 import com.pmatuki.wowpapers.remote.download.ImageDownloadService
+import com.pmatuki.wowpapers.remote.download.ImageDownloadServiceImpl
 import com.pmatuki.wowpapers.view.mapper.WallpaperMapper
 import toothpick.Scope
 import toothpick.config.Module
@@ -20,7 +22,7 @@ object WallpaperListViewModelScope {
 
 object WallpaperListViewModule : Module() {
     init {
-        bind<WallpaperDataSource>()
+        bind<WallpaperDataSource>().toClass<WallpaperDataSourceImpl>()
         bind<WallpaperMapper>()
     }
 }
@@ -33,7 +35,7 @@ object DetailViewModelScope {
 
 object DetailViewModelModule : Module() {
     init {
-        bind<ImageDownloadService>()
-        bind<WallpaperService>()
+        bind<ImageDownloadService>().toClass<ImageDownloadServiceImpl>()
+        bind<WallpaperService>().toClass<WallpaperServiceImpl>()
     }
 }
