+++ app/src/main/java/com/pmatuki/wowpapers/view/list/WallpaperListViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.pmatuki.wowpapers.remote.WallpaperDataSource
-import com.pmatuki.wowpapers.view.di.DetailViewModelScope
+import com.pmatuki.wowpapers.remote.api.WallpaperDataSource
+import com.pmatuki.wowpapers.remote.api.WallpaperDataSourceImpl
 import com.pmatuki.wowpapers.view.di.WallpaperListViewModelScope
 import com.pmatuki.wowpapers.view.mapper.WallpaperMapper
 import kotlinx.coroutines.launch
